# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: EV Community Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /owners:
    get:
      operationId: listOwners
      tags: [owners]
      summary: List all approved EV owners
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: carModel
          in: query
          schema:
            type: string
        - name: city
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of owners
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Owner"
  /owners/{id}:
    get:
      operationId: getOwner
      tags: [owners]
      summary: Get a specific owner's profile
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Owner profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Owner"
        "404":
          description: Owner not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /owners/{id}/profile:
    patch:
      operationId: updateOwnerProfile
      tags: [owners]
      summary: Update owner profile (self only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateOwnerBody"
      responses:
        "200":
          description: Updated owner
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Owner"

  # ─── ADMIN ──────────────────────────────────────────────────────────────────
...
  /range/calculate:
    post:
      operationId: calculateRange
      tags: [range]
      summary: Calculate estimated range based on conditions
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RangeCalculateBody"
      responses:
        "200":
          description: Range calculation result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RangeResult"

  # ─── CPO ─────────────────────────────────────────────────────────────────────
  /cpo:
    get:
      operationId: listCpos
      tags: [cpo]
      summary: List Charge Point Operators
      parameters:
        - name: city
          in: query
          schema:
            type: string
        - name: network
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of CPOs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Cpo"

    post:
      operationId: createCpo
      tags: [cpo]
      summary: Add a CPO (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCpoBody"
      responses:
        "201":
          description: CPO created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cpo"
  /cpo/{id}:
    patch:
      operationId: updateCpo
      tags: [cpo]
      summary: Update a CPO (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateCpoBody"
      responses:
        "200":
          description: CPO updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cpo"

    delete:
      operationId: deleteCpo
      tags: [cpo]
      summary: Delete a CPO (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: CPO deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"

  # ─── ACCESSORIES ─────────────────────────────────────────────────────────────
...
  /gallery:
    get:
      operationId: listGalleryItems
      tags: [gallery]
      summary: List gallery photos
      parameters:
        - name: ownerId
          in: query
          schema:
            type: integer
        - name: carModel
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of gallery items
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/GalleryItem"

    post:
      operationId: createGalleryItem
      tags: [gallery]
      summary: Upload a gallery photo (approved owners)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateGalleryItemBody"
      responses:
        "201":
          description: Gallery item created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GalleryItem"
  /gallery/{id}:
    delete:
      operationId: deleteGalleryItem
      tags: [gallery]
      summary: Delete a gallery item
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Gallery item deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"

  # ─── FAQ / AI ─────────────────────────────────────────────────────────────────
  /faq/questions:
    get:
      operationId: listFaqQuestions
      tags: [faq]
      summary: List common FAQ questions and answers
      parameters:
        - name: carModel
          in: query
          schema:
            type: string
        - name: category
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of FAQ items
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/FaqItem"

    post:
      operationId: createFaqQuestion
      tags: [faq]
      summary: Add FAQ knowledge base entry (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateFaqItemBody"
      responses:
        "201":
          description: FAQ item created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FaqItem"
  /faq/ask:
    post:
      operationId: askFaq
      tags: [faq]
      summary: Ask the AI assistant a question about EVs
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AskFaqBody"
      responses:
        "200":
          description: AI response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FaqAnswer"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required: [error]
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required: [message]

    # ─── AUTH ─────────────────────────────────────────────────────────────────
...
    Owner:
      type: object
      properties:
        id:
          type: integer
        email:
          type: string
        name:
          type: string
        age:
          type: integer
        phoneNumber:
          type: string
        whatsappNumber:
          type: string
        telegramId:
          type: string
        areaOfStay:
          type: string
        city:
          type: string
        occupation:
          type: string
        vehicleNumber:
          type: string
        evCarsOwned:
          type: array
          items:
            type: string
        variantColor:
          type: string
        otherCarModel:
          type: string
        purchaseMonthYear:
          type: string
        status:
          type: string
          enum: [pending, approved, rejected]
        role:
          type: string
          enum: [user, admin]
        blogEnabled:
          type: boolean
        username:
          type: string
        avatarUrl:
          type: string
        bio:
          type: string
        createdAt:
          type: string
          format: date-time
      required: [id, email, name, status, role]
    UpdateOwnerBody:
      type: object
      properties:
        bio:
          type: string
        avatarUrl:
          type: string
        whatsappNumber:
          type: string
        telegramId:
          type: string
...
    RangeCalculateBody:
      type: object
      properties:
        carCode:
          type: string
        batteryPercentage:
          type: number
          minimum: 0
          maximum: 100
        drivingStyle:
          type: string
          enum: [eco, normal, sport]
        acUsage:
          type: string
          enum: [off, low, high]
        terrainType:
          type: string
          enum: [flat, hilly, highway]
        ambientTempCelsius:
          type: number
      required: [carCode, batteryPercentage, drivingStyle, acUsage, terrainType]
    RangeResult:
      type: object
      properties:
        estimatedRangeKm:
          type: number
        carName:
          type: string
        batteryUsable:
          type: number
        factors:
          type: object
          properties:
            drivingStyleFactor:
              type: number
            acFactor:
              type: number
            terrainFactor:
              type: number
            tempFactor:
              type: number
          required: [drivingStyleFactor, acFactor, terrainFactor, tempFactor]
        tips:
          type: array
          items:
            type: string
      required: [estimatedRangeKm, carName, batteryUsable, factors, tips]

    # ─── CPO ──────────────────────────────────────────────────────────────────
    Cpo:
      type: object
      properties:
        id:
          type: integer
        networkName:
          type: string
        contactPerson:
          type: string
        email:
          type: string
        phone:
          type: string
        whatsapp:
          type: string
        city:
          type: string
        state:
          type: string
        address:
          type: string
        chargerTypes:
          type: array
          items:
            type: string
        maxKw:
          type: number
        operatingHours:
          type: string
        notes:
          type: string
        logoUrl:
          type: string
        createdAt:
          type: string
          format: date-time
      required: [id, networkName, city, state]
    CreateCpoBody:
      type: object
      properties:
        networkName:
          type: string
        contactPerson:
          type: string
        email:
          type: string
        phone:
          type: string
        whatsapp:
          type: string
        city:
          type: string
        state:
          type: string
        address:
          type: string
        chargerTypes:
          type: array
          items:
            type: string
        maxKw:
          type: number
        operatingHours:
          type: string
        notes:
          type: string
        logoUrl:
          type: string
      required: [networkName, city, state]
    UpdateCpoBody:
      type: object
      properties:
        networkName:
          type: string
        contactPerson:
          type: string
        email:
          type: string
        phone:
          type: string
        whatsapp:
          type: string
        city:
          type: string
        state:
          type: string
        address:
          type: string
        chargerTypes:
          type: array
          items:
            type: string
        maxKw:
          type: number
        operatingHours:
          type: string
        notes:
          type: string
        logoUrl:
          type: string

    # ─── ACCESSORIES ──────────────────────────────────────────────────────────
...
    GalleryItem:
      type: object
      properties:
        id:
          type: integer
        imageUrl:
          type: string
        caption:
          type: string
        carModel:
          type: string
        location:
          type: string
        ownerId:
          type: integer
        ownerName:
          type: string
        ownerAvatarUrl:
          type: string
        likeCount:
          type: integer
        createdAt:
          type: string
          format: date-time
      required: [id, imageUrl, ownerId, ownerName, createdAt]
    CreateGalleryItemBody:
      type: object
      properties:
        imageUrl:
          type: string
        caption:
          type: string
        carModel:
          type: string
        location:
          type: string
      required: [imageUrl]

    # ─── FAQ ──────────────────────────────────────────────────────────────────
    FaqItem:
      type: object
      properties:
        id:
          type: integer
        question:
          type: string
        answer:
          type: string
        category:
          type: string
        carModels:
          type: array
          items:
            type: string
        helpful:
          type: integer
        createdAt:
          type: string
          format: date-time
      required: [id, question, answer, category]
    CreateFaqItemBody:
      type: object
      properties:
        question:
          type: string
        answer:
          type: string
        category:
          type: string
        carModels:
          type: array
          items:
            type: string
      required: [question, answer, category]
    AskFaqBody:
      type: object
      properties:
        question:
          type: string
        carModel:
          type: string
        conversationId:
          type: string
      required: [question]
    FaqAnswer:
      type: object
      properties:
        answer:
          type: string
        conversationId:
          type: string
        relatedFaqs:
          type: array
          items:
            $ref: "#/components/schemas/FaqItem"
        sources:
          type: array
          items:
            type: string
      required: [answer, conversationId]
...